Refactored property stuff to data::property
