+++ src/data/mod.rs
@@ -1,5 +1,9 @@
 mod attribute;
+mod property;
 
-pub use attribute::AttributeValue;
 pub use attribute::get_attribute;
 pub use attribute::store_metadata;
+pub use attribute::AttributeValue;
+
+pub use property::get_property;
+pub use property::Path as PropertyPath;
